+++ appmap/test/data/unittest/simple/__init__.py
@@ -7,3 +7,9 @@ class Simple:
 
     def hello_world(self, bang):
         return '%s %s%s' % (self.hello(), self.world(), bang)
+
+    def getReady(self):
+        pass
+
+    def finishUp(self):
+        pass