+++ appmap/unittest.py
@@ -1,9 +1,10 @@
+from contextlib import contextmanager
 import unittest
 
 from appmap._implementation import testing_framework
+import appmap.wrapt as wrapt
 
 
-original_run = unittest.TestCase.run
 session = testing_framework.session('unittest')
 
 def get_test_location(cls, method_name):
@@ -12,23 +13,29 @@ def get_test_location(cls, method_name):
     return get_function_location(fn)
 
 
-def run(self, result=None):
-    method_name = self.id().split('.')[-1]
-    # Use the result's location if provided (e.g. by pytest),
-    # otherwise cobble one together ourselves.
-    if hasattr(result, 'location'):
-        location = result.location
-    else:
-        location = get_test_location(self.__class__, method_name)
+# unittest.case._Outcom.testPartExecutor is used by all supported
+# versions of unittest to run test cases. `isTest` will be True when
+# the part is the actual test method, False when it's setUp or
+# teardown.
+@wrapt.patch_function_wrapper('unittest.case', '_Outcome.testPartExecutor')
+@contextmanager
+def testPartExecutor(wrapped, _, args, kwargs):
+    def _args(test_case, *_, isTest=False, **__):
+        return (test_case, isTest)
+    test_case, is_test = _args(*args, **kwargs)
+    if not is_test:
+        with wrapped(*args, **kwargs):
+            yield
+        return
+
+    method_name = test_case.id().split('.')[-1]
+    location = get_test_location(test_case.__class__, method_name)
     with session.record(
-            self.__class__,
+            test_case.__class__,
             method_name,
             location=location):
-        original_run(self, result)
-
-
-unittest.TestCase.run = run
-
+        with wrapped(*args, **kwargs):
+            yield
 
 if __name__ == '__main__':
     unittest.main(module=None)